% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribuidores_tjrj.R
\docType{data}
\name{distribuidores_tjrj}
\alias{distribuidores_tjrj}
\title{distribuidores_tjrj}
\format{
tibble com 108 obs e duas colunas:
\describe{
\item{\code{codigo}}{código do distribuidor}
\item{\code{distribuidor}}{distribuidor}
}
}
\source{
\url{https://tjrj.jus.br/}
}
\usage{
distribuidores_tjrj
}
\description{
Códigos dos distribuidores do TJRJ
}
\keyword{datasets}
