% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_movimentacao_tjrj.R
\name{ler_movimentacao_tjrj}
\alias{ler_movimentacao_tjrj}
\title{Ler movimentação do tjrj}
\usage{
ler_movimentacao_tjrj(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório se arquivos não forem informados}
}
\value{
tibble
}
\description{
Ler movimentação do tjrj
}
