% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjrj_baixar_cpopg.R
\name{tjrj_baixar_cpopg}
\alias{tjrj_baixar_cpopg}
\title{Baixar cpopg do TJRJ}
\usage{
tjrj_baixar_cpopg(processo = NULL, diretorio = ".")
}
\arguments{
\item{processo}{Número do processo com ou sem separador}

\item{diretorio}{Diretório onde serão armazenados os htmls}
}
\value{
htmls baixados
}
\description{
Baixar cpopg do TJRJ
}
