% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjrj_ler_dados.R
\name{tjrj_ler_dados}
\alias{tjrj_ler_dados}
\title{Ler dados do TJRJ}
\usage{
tjrj_ler_dados(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se não informar os arquivos}
}
\value{
tibble
}
\description{
Ler dados do TJRJ
}
